examples/movie_bot/requirements.txt
mirascope
openai
python-dotenv
